backend/requirements.txt
flask==2.3.3
werkzeug==2.3.7
anthropic
python-dotenv==1.0.0
gunicorn==21.2.0
flask-cors==5.0.1
cartesia
pyaudio